Make Carousel more responsive on mobile
Currently on mobile the ArticleCards don't become narrower than 350px,
which together with the page padding (20px * 2) and borders (1px * 2),
means that the Carousel can't adapt to screens with a viewport narrower
than 392px.

To make the Carousel with ArticleCards truly responsive, the fixed
height is replaced with a bottom padding and the width is calculated
based on the viewport width (100vw) minus the page padding and borders.