Add error handling to fetch & enable client use
This adds various error handling to the fetchApi/Interceptor, such as
throwing errors if the server returns an error and showing a toaster
message with the error. However, client errors are just thrown and need
to be handled on a case-by-case basis, depending on how the error should
be presented to the user.

This also enables using the `fetchApi` in client components, as the
isomorphic cookie utility is used.