Check if token expired to improve logged in check
Don't just check that if there is a JWT token, but also check if the
token has expired and in that case show a Toast that the login has
expired.

Note that it's very important to not rely on `mightBeLoggedIn` fully.
Only the backend can validate the token signature to guarantee if the
user is authenticated correctly. The function `mightBeLoggedIn` can only
serve as a hint that the user is probably _not_ logged in. Even if it
returns `true` the API may still return a `401 Unauthorized`.