Unify chat APIs & add error handling
Similar code was used in the components `<SuggestedPrompt>´,
`<NewChat>`, `<Chatbot>` and `<AskMaestro>` components for starting a
chat and asking the chatbot a question. This code has been unified into
a single chat API utility, which is then used in all these components.
Only the redirecting logic is left in each component, as it is a view
layer concern that requires access to the router.

In addition to the existing `fetchInterceptor` error handling, this
handles client errors by showing a toast with the error message returned
by the server. This also ensure to guard for the case of there not being
any chat ID and in that case to not redirect the user to
`/my/chats/undefined`.