Unify chat prompts
Rather than having 3 similar chat prompt input components, this unifies
`<NewChat>`, `<AskMaestro>` and `<Chatbot>` into a single `<ChatPrompt>`
component that is capable of both starting a new chat, asking a question
in an existing chat or both.

In addition to less code and consistent styling, this brings the benefit
of all 3 places sharing the loading indicator.