Improve DesktopNavLinks inverted rounded corners
The overlayed circle that melt into the background color to make the
inverted rounded white corners between the navigation link/item and the
page below, are larger than they should be, resulting in a jagged line
and risk overlaying the adjacent menu items.

This decreases the size of these circles to 20px, meaning a radius of
10px that match the 10px size of the white square. This results in a
smoother rounded corner.