Prevent own DDoS attack on the API
In the mobile view, the `MobileChatsMenu` fetches list of the chats in
a `useEffect` hook. As the `useEffect` hook doesn't have any dependency
array, it will run on each render. Each render will trigger a new fetch
and each fetch sets the `chats` state, which triggers a new render.

This causes an infinite loop that calls the `/my/chats` API. In a
production scenario, several mobile visitors to the `/my/chats` page
could basically cause a Distributed Denial of Service attack against the
API by unsuspecting, legitimate users.

This creates a `useOnMount` hook that will only run once per component
mount. While super simple, it introduces a reusable concept that can
prevent this issue in the future.