Use more optimized JPG file on login page
The `login.jpg` is 65% smaller than the PNG equivalent (60 kB instead of
171 kB), so the JPG is preferred for performance and the quality
difference should not be noticeable by the average user.

Note that the PNG is removed here, as it is no longer used. It will
forever remain in the git history though, where it can be retrieved with
a `git checkout 5516c8d -- ./public/login.png`