Show help menu & simplify roles check
The issue is that the "Help" url/route is not referenced in the
`isUserAllowed` utility. However, rather than just adding it, let's use
the `for` property of the links to determine a user's permissions to
access a specific page.

The `for` property of the links is updated to match the `isUserAllowed`
utility, while the `isUserAllowed` utility is made to handle not only
single user role matching, but allowing matching several user roles
against several permissible roles (`allowedRoles`) in addition to
accepting a single string role or `all` that permits all users to access
the URL.