Verify JWT expiry in middleware
So far the fact that a JWT cookie existed, meant that the middleware
assumed the user is logged in. This ensures the JWT cookie is in fact
a JWT token, that it can be base64 decoded and parsed as JSON and that
it has an expiry date that has not passed.

However, this still does NOT validate the JWT signature - only the
backend can do that and it is generally not recommended to do API calls
in a middleware. In the end the only way to know if the user is logged
in with a valid JWT, is to make an authenticated API call and see if it
returns successfully or with a `401 Unauthorized` HTTP error status.