Create AuthenticatedPage HOC page wrapper
While the middleware checks for the existence of the JWT cookie and if
it is a parseable JWT token with a valid expiry date, the frontend can
never guarantee the JWT validity. The backend API may still return a
`401 Unauthorized`.

We can't check authentication or authorization in the middleware nor the
layout, as asynchronous and data fetching logic should be avoided in the
middleware and the layout does not run on every page navigation, only
first full page load, not subsequent single page navigations.

So this creates a Higher-Order Component that can wrap any page export
to ensure the user is logged in (authenticated) and has the correct
authorization permissions to access a specific page.