Redirect logged in users on login page
If the user has a seemingly valid JWT token, redirect to the `/home`
page to avoid forcing the user to login again just because they came
to the `/login` page.

If the user for some, unusual reason has an invalid JWT token, such as
invalid signature that only the backend can validate, the user will get
redirected back to the `/login` in page in `<AuthenticatedPage>` with
the `query parameter `?unauthorized=true` to ensure the user does not
end up in a redirect loop.