Fix broken styling in mobile nav menu
When using the Radix Dialog, it creates a Portal, which is placed
outside the normal react mounting point and therefore also outside the
root layout `<Theme>`.

This wraps the dialog content in the same default theme as the root
layout. This particularly fixes the `<AvatarBox>` styling, that was
completely lacking.